requirements.txt
# python_version 3.10
# pip_version 24.*

ExifRead==3.0.0
matplotlib==3.9.2
numpy==2.1.1
opencv_python==4.7.0.72
Pillow==10.4.0
pytest==8.2.2
python-dotenv==1.0.1
PyYAML==6.0.2
scipy==1.14.1
scikit-image==0.24.0
